--- a/src/main/java/org/springframework/data/convert/CustomConversions.java
+++ b/src/main/java/org/springframework/data/convert/CustomConversions.java
@@ -393,7 +393,10 @@ public class CustomConversions {
 		public Class<?> computeIfAbsent(Class<?> sourceType, Class<?> targetType,
 				Function<ConvertiblePair, Class<?>> mappingFunction) {
 
-			TargetTypes targetTypes = customReadTargetTypes.computeIfAbsent(sourceType, TargetTypes::new);
+			TargetTypes targetTypes = customReadTargetTypes.get(sourceType);
+			if (targetTypes == null) {
+				targetTypes = customReadTargetTypes.computeIfAbsent(sourceType, TargetTypes::new);
+			}
 			return targetTypes.computeIfAbsent(targetType, mappingFunction);
 		}